add sudoku

+++ includes/functions_users_delete.php
@@ -119,11 +119,14 @@ function ip_user_kill($user_id)
 		$sql = "DELETE FROM " . DL_NOTRAF_TABLE . " WHERE user_id = " . $user_id;
 		$db->sql_query($sql);
 
-		$sql = "DELETE FROM " . SUDOKU_STATS . " WHERE user_id = " . $user_id;
-		$db->sql_query($sql);
+		if (!empty($config['plugins']['sudoku']['enabled']))
+		{
+			$sql = "DELETE FROM " . SUDOKU_STATS . " WHERE user_id = " . $user_id;
+			$db->sql_query($sql);
 
-		$sql = "DELETE FROM " . SUDOKU_USERS . " WHERE user_id = " . $user_id;
-		$db->sql_query($sql);
+			$sql = "DELETE FROM " . SUDOKU_USERS . " WHERE user_id = " . $user_id;
+			$db->sql_query($sql);
+		}
 
 		// Start add - Fully integrated shoutbox MOD
 		$sql = "UPDATE " . SHOUTBOX_TABLE . "
